@@ -1,34 +1,63 @@
-module github.com/micro/micro/v2
+module github.com/micro/micro/v3
 
-go 1.13
+go 1.15
 
 require (
-	github.com/aws/aws-sdk-go v1.23.0
+	github.com/bitly/go-simplejson v0.5.0
+	github.com/blang/semver v3.5.1+incompatible
+	github.com/bmizerany/assert v0.0.0-20160611221934-b7ed37b82869 // indirect
 	github.com/boltdb/bolt v1.3.1
+	github.com/caddyserver/certmagic v0.10.6
 	github.com/chzyer/logex v1.1.10 // indirect
 	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e
 	github.com/chzyer/test v0.0.0-20180213035817-a1ea475d72b1 // indirect
-	github.com/cloudflare/cloudflare-go v0.10.9
+	github.com/cloudflare/cloudflare-go v0.10.9 // indirect
+	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
+	github.com/davecgh/go-spew v1.1.1
+	github.com/desertbit/timer v0.0.0-20180107155436-c41aec40b27f // indirect
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/dustin/go-humanize v1.0.0
-	github.com/fsnotify/fsnotify v1.4.7
+	github.com/evanphx/json-patch/v5 v5.0.0
+	github.com/fsnotify/fsnotify v1.4.9 // indirect
+	github.com/getkin/kin-openapi v0.26.0
 	github.com/go-acme/lego/v3 v3.4.0
-	github.com/golang/protobuf v1.4.2
-	github.com/google/uuid v1.1.1
+	github.com/gobwas/httphead v0.0.0-20180130184737-2c6c146eadee
+	github.com/gobwas/pool v0.2.1 // indirect
+	github.com/gobwas/ws v1.0.3
+	github.com/golang/protobuf v1.4.3
+	github.com/google/uuid v1.1.2
 	github.com/gorilla/handlers v1.4.2
 	github.com/gorilla/mux v1.7.3
-	github.com/micro/cli/v2 v2.1.2
-	github.com/micro/go-micro/v2 v2.8.1-0.20200603084508-7b379bf1f16e
+	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/hashicorp/go-version v1.2.1
+	github.com/hpcloud/tail v1.0.0
+	github.com/improbable-eng/grpc-web v0.13.0
+	github.com/juju/fslock v0.0.0-20160525022230-4d5c94c67b4b
+	github.com/kr/pretty v0.2.0
+	github.com/kr/text v0.2.0 // indirect
 	github.com/miekg/dns v1.1.27
-	github.com/netdata/go-orchestrator v0.0.0-20190905093727-c793edba0e8f
+	github.com/minio/minio-go/v7 v7.0.5
+	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/olekukonko/tablewriter v0.0.4
+	github.com/onsi/ginkgo v1.12.0 // indirect
+	github.com/onsi/gomega v1.7.1
+	github.com/oxtoacart/bpool v0.0.0-20190530202638-03653db5a59c
 	github.com/patrickmn/go-cache v2.1.0+incompatible
 	github.com/pkg/errors v0.9.1
+	github.com/rhysd/go-github-selfupdate v1.2.2
+	github.com/rs/cors v1.7.0 // indirect
 	github.com/serenize/snaker v0.0.0-20171204205717-a683aaf2d516
-	github.com/spf13/viper v1.6.3
+	github.com/stretchr/testify v1.6.1
+	github.com/teris-io/shortid v0.0.0-20171029131806-771a37caa5cf
+	github.com/urfave/cli/v2 v2.2.0
+	github.com/xanzy/go-gitlab v0.35.1
 	github.com/xlab/treeprint v0.0.0-20181112141820-a009c3971eca
-	golang.org/x/crypto v0.0.0-20200510223506-06a226fb4e37
-	golang.org/x/net v0.0.0-20200520182314-0ba52f642ac2
-	golang.org/x/tools v0.0.0-20191216173652-a0e659d51361
-	google.golang.org/genproto v0.0.0-20191216164720-4f79533eabd1
-	google.golang.org/grpc v1.26.0
+	go.etcd.io/bbolt v1.3.5
+	golang.org/x/crypto v0.0.0-20200709230013-948cd5f35899
+	golang.org/x/net v0.0.0-20200707034311-ab3426394381
+	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013
+	google.golang.org/grpc v1.27.0
+	google.golang.org/protobuf v1.25.0
+	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
+	gopkg.in/square/go-jose.v2 v2.4.1 // indirect
 )
