@@ -5,57 +5,71 @@ package util
 import (
 	"encoding/json"
 	"fmt"
-	"net"
-	"os"
+	"regexp"
 	"strings"
 
-	ccli "github.com/micro/cli/v2"
-	"github.com/micro/micro/v2/internal/config"
-	"github.com/micro/micro/v2/internal/platform"
-	"github.com/micro/micro/v2/service/runtime/profile"
+	merrors "github.com/micro/micro/v3/service/errors"
+
+	"github.com/micro/micro/v3/internal/config"
+	"github.com/urfave/cli/v2"
 )
 
 const (
-	// EnvLocal is a builtin environment, it means services launched
-	// with `micro run` will use default, zero dependency implementations for
-	// interfaces, like mdns for registry.
+	// EnvLocal is a builtin environment, it represents your local `micro server`
 	EnvLocal = "local"
-	// EnvServer is a builtin environment, it represents your local `micro server`
-	EnvServer = "server"
-	// EnvPlatform is a builtin environment, the One True Micro Live(tm) environment.
+	// EnvDev is a builtin staging / dev environment in the cloud
+	EnvDev = "dev"
+	// EnvPlatform is a builtin highly available environment in the cloud,
 	EnvPlatform = "platform"
 )
 
 const (
-	// localProxyAddress is the default proxy address for environment local
-	// local env does not use other services so talking about a proxy
-	localProxyAddress = "none"
-	// serverProxyAddress is the default proxy address for environment server
-	serverProxyAddress = "127.0.0.1:8081"
-	// platformProxyAddress is teh default proxy address for environment platform
-	platformProxyAddress = "proxy.micro.mu"
+	// localProxyAddress is the default proxy address for environment server
+	localProxyAddress = "127.0.0.1:8081"
+	// deprecated dev env
+	devProxyAddress = "proxy.m3o.dev"
+	// platformProxyAddress is the default proxy address for environment platform
+	platformProxyAddress = "proxy.m3o.com"
+)
+
+var (
+	// list of services managed
+	// TODO: make use server/server list
+	services = []string{
+		// runtime services
+		"network",  // :8085 (peer), :8443 (proxy)
+		"runtime",  // :8088
+		"registry", // :8000
+		"config",   // :8001
+		"store",    // :8002
+		"broker",   // :8003
+		"router",   // :8084
+		"auth",     // :8010
+		"proxy",    // :8081
+		"api",      // :8080
+	}
 )
 
 var defaultEnvs = map[string]Env{
-	EnvLocal: Env{
+	EnvLocal: {
 		Name:         EnvLocal,
 		ProxyAddress: localProxyAddress,
+		Description:  "Local running Micro Server",
 	},
-	EnvServer: Env{
-		Name:         EnvServer,
-		ProxyAddress: serverProxyAddress,
+	EnvDev: {
+		Name:         EnvDev,
+		ProxyAddress: devProxyAddress,
+		Description:  "Deprecated: Please use platform environment",
 	},
-	EnvPlatform: Env{
+	EnvPlatform: {
 		Name:         EnvPlatform,
 		ProxyAddress: platformProxyAddress,
+		Description:  "Cloud hosted Micro Platform",
 	},
 }
 
-func isBuiltinService(command string) bool {
-	if command == "server" {
-		return true
-	}
-	for _, service := range platform.Services {
+func IsBuiltInService(command string) bool {
+	for _, service := range services {
 		if command == service {
 			return true
 		}
@@ -63,106 +77,85 @@ func isBuiltinService(command string) bool {
 	return false
 }
 
-// SetupCommand includes things that should run for each command.
-func SetupCommand(ctx *ccli.Context) {
-	if ctx.Args().Len() == 1 && isBuiltinService(ctx.Args().First()) {
-		return
-	}
-	if ctx.Args().Len() >= 1 && ctx.Args().First() == "env" {
-		return
+// CLIProxyAddress returns the proxy address which should be set for the client
+func CLIProxyAddress(ctx *cli.Context) (string, error) {
+	switch ctx.Args().First() {
+	case "new", "server", "help", "env":
+		return "", nil
 	}
 
-	toFlag := func(s string) string {
-		return strings.ToLower(strings.ReplaceAll(s, "MICRO_", ""))
-	}
-	setFlags := func(envars []string) {
-		for _, envar := range envars {
-			// setting both env and flags here
-			// as the proxy settings for example did not take effect
-			// with only flags
-			parts := strings.Split(envar, "=")
-			key := toFlag(parts[0])
-			os.Setenv(parts[0], parts[1])
-			ctx.Set(key, parts[1])
-		}
+	// fix for "micro service [command]", e.g "micro service auth"
+	if ctx.Args().First() == "service" && IsBuiltInService(ctx.Args().Get(1)) {
+		return "", nil
 	}
 
-	env := GetEnv(ctx)
-
-	// if we're running a local environment return here
-	if len(env.ProxyAddress) == 0 || env.Name == EnvLocal {
-		return
+	// don't set the proxy address on the proxy
+	if ctx.Args().First() == "proxy" {
+		return "", nil
 	}
 
-	switch env.Name {
-	case EnvServer:
-		setFlags(profile.ServerCLI())
-	case EnvPlatform:
-		setFlags(profile.PlatformCLI())
-	default:
-		// default case for ad hoc envs, see comments above about tests
-		setFlags(profile.ServerCLI())
+	env, err := GetEnv(ctx)
+	if err != nil {
+		return "", err
 	}
-
-	// Set the proxy
-	setFlags([]string{"MICRO_PROXY=" + env.ProxyAddress})
+	addr := env.ProxyAddress
+	if !strings.Contains(addr, ":") {
+		return fmt.Sprintf("%v:443", addr), nil
+	}
+	return addr, nil
 }
 
 type Env struct {
 	Name         string
 	ProxyAddress string
+	Description  string
 }
 
-func AddEnv(env Env) {
-	envs := getEnvs()
+func AddEnv(env Env) error {
+	envs, err := getEnvs()
+	if err != nil {
+		return err
+	}
 	envs[env.Name] = env
-	setEnvs(envs)
+	return setEnvs(envs)
 }
 
-func getEnvs() map[string]Env {
+func getEnvs() (map[string]Env, error) {
 	envsJSON, err := config.Get("envs")
 	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
+		return nil, fmt.Errorf("Error getting environment: %v", err)
 	}
 	envs := map[string]Env{}
 	if len(envsJSON) > 0 {
 		err := json.Unmarshal([]byte(envsJSON), &envs)
 		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
+			return nil, err
 		}
 	}
 	for k, v := range defaultEnvs {
 		envs[k] = v
 	}
-	return envs
+	return envs, nil
 }
 
-func setEnvs(envs map[string]Env) {
+func setEnvs(envs map[string]Env) error {
 	envsJSON, err := json.Marshal(envs)
 	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
+		return err
 	}
-	err = config.Set(string(envsJSON), "envs")
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
+	return config.Set("envs", string(envsJSON))
 }
 
 // GetEnv returns the current selected environment
 // Does not take
-func GetEnv(ctx *ccli.Context) Env {
+func GetEnv(ctx *cli.Context) (Env, error) {
 	var envName string
 	if len(ctx.String("env")) > 0 {
 		envName = ctx.String("env")
 	} else {
 		env, err := config.Get("env")
 		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
+			return Env{}, err
 		}
 		if env == "" {
 			env = EnvLocal
@@ -173,60 +166,136 @@ func GetEnv(ctx *ccli.Context) Env {
 	return GetEnvByName(envName)
 }
 
-func GetEnvByName(env string) Env {
-	envs := getEnvs()
-
+func GetEnvByName(env string) (Env, error) {
+	envs, err := getEnvs()
+	if err != nil {
+		return Env{}, err
+	}
 	envir, ok := envs[env]
 	if !ok {
-		fmt.Println(fmt.Sprintf("Env \"%s\" not found. See `micro env` for available environments.", env))
-		os.Exit(1)
+		return Env{}, fmt.Errorf("Env \"%s\" not found. See `micro env` for available environments.", env)
 	}
-
-	if len(envir.ProxyAddress) == 0 {
-		return envir
-	}
-
-	// default to :443
-	if _, port, _ := net.SplitHostPort(envir.ProxyAddress); len(port) == 0 {
-		envir.ProxyAddress = net.JoinHostPort(envir.ProxyAddress, "443")
-	}
-
-	return envir
+	return envir, nil
 }
 
-func GetEnvs() []Env {
-	envs := getEnvs()
-	ret := []Env{defaultEnvs[EnvLocal], defaultEnvs[EnvServer], defaultEnvs[EnvPlatform]}
-	nonDefaults := []Env{}
+func GetEnvs() ([]Env, error) {
+	envs, err := getEnvs()
+	if err != nil {
+		return nil, err
+	}
+
+	var ret []Env
+
+	// populate the default environments
+	for _, env := range defaultEnvs {
+		ret = append(ret, env)
+	}
+
+	var nonDefaults []Env
+
 	for _, env := range envs {
 		if _, isDefault := defaultEnvs[env.Name]; !isDefault {
 			nonDefaults = append(nonDefaults, env)
 		}
 	}
+
 	// @todo order nondefault envs alphabetically
 	ret = append(ret, nonDefaults...)
-	return ret
+
+	return ret, nil
 }
 
 // SetEnv selects an environment to be used.
-func SetEnv(envName string) {
-	envs := getEnvs()
+func SetEnv(envName string) error {
+	envs, err := getEnvs()
+	if err != nil {
+		return err
+	}
 	_, ok := envs[envName]
 	if !ok {
-		fmt.Printf("Environment '%v' does not exist\n", envName)
-		os.Exit(1)
+		return fmt.Errorf("Environment '%v' does not exist", envName)
 	}
-	config.Set(envName, "env")
+	return config.Set("env", envName)
 }
 
-func IsLocal(ctx *ccli.Context) bool {
-	return GetEnv(ctx).Name == EnvLocal
+// DelEnv deletes an env from config
+func DelEnv(envName string) error {
+	envs, err := getEnvs()
+	if err != nil {
+		return err
+	}
+	_, ok := envs[envName]
+	if !ok {
+		return fmt.Errorf("Environment '%v' does not exist", envName)
+	}
+	delete(envs, envName)
+	return setEnvs(envs)
 }
 
-func IsServer(ctx *ccli.Context) bool {
-	return GetEnv(ctx).Name == EnvServer
+func IsPlatform(ctx *cli.Context) bool {
+	env, err := GetEnv(ctx)
+	if err == nil && env.Name == EnvPlatform {
+		return true
+	}
+	return false
 }
 
-func IsPlatform(ctx *ccli.Context) bool {
-	return GetEnv(ctx).Name == EnvPlatform
+type Exec func(*cli.Context, []string) ([]byte, error)
+
+func Print(e Exec) func(*cli.Context) error {
+	return func(c *cli.Context) error {
+		rsp, err := e(c, c.Args().Slice())
+		if err != nil {
+			return CliError(err)
+		}
+		if len(rsp) > 0 {
+			fmt.Printf("%s\n", string(rsp))
+		}
+		return nil
+	}
+}
+
+// CliError returns a user friendly message from error. If we can't determine a good one returns an error with code 128
+func CliError(err error) cli.ExitCoder {
+	if err == nil {
+		return nil
+	}
+	// if it's already a cli.ExitCoder we use this
+	cerr, ok := err.(cli.ExitCoder)
+	if ok {
+		return cerr
+	}
+
+	// grpc errors
+	if mname := regexp.MustCompile(`malformed method name: \\?"(\w+)\\?"`).FindStringSubmatch(err.Error()); len(mname) > 0 {
+		return cli.Exit(fmt.Sprintf(`Method name "%s" invalid format. Expecting service.endpoint`, mname[1]), 3)
+	}
+	if service := regexp.MustCompile(`service ([\w\.]+): route not found`).FindStringSubmatch(err.Error()); len(service) > 0 {
+		return cli.Exit(fmt.Sprintf(`Service "%s" not found`, service[1]), 4)
+	}
+	if service := regexp.MustCompile(`unknown service ([\w\.]+)`).FindStringSubmatch(err.Error()); len(service) > 0 {
+		if strings.Contains(service[0], ".") {
+			return cli.Exit(fmt.Sprintf(`Service method "%s" not found`, service[1]), 5)
+		}
+		return cli.Exit(fmt.Sprintf(`Service "%s" not found`, service[1]), 5)
+	}
+	if address := regexp.MustCompile(`Error while dialing dial tcp.*?([\w]+\.[\w:\.]+): `).FindStringSubmatch(err.Error()); len(address) > 0 {
+		return cli.Exit(fmt.Sprintf(`Failed to connect to micro server at %s`, address[1]), 4)
+	}
+
+	merr, ok := err.(*merrors.Error)
+	if !ok {
+		return cli.Exit(err, 128)
+	}
+
+	switch merr.Code {
+	case 408:
+		return cli.Exit("Request timed out", 1)
+	case 401:
+		// TODO check if not signed in, prompt to sign in
+		return cli.Exit("Not authorized to perform this request", 2)
+	}
+
+	// fallback to using the detail from the merr
+	return cli.Exit(merr.Detail, 127)
 }
