@@ -1,20 +1,22 @@
 // Code generated by protoc-gen-micro. DO NOT EDIT.
-// source: examples/greeter/greeter.proto
+// source: cmd/protoc-gen-micro/examples/greeter/greeter.proto
 
 package greeter
 
 import (
 	fmt "fmt"
+	math "math"
+
 	proto "github.com/golang/protobuf/proto"
 	_ "google.golang.org/genproto/googleapis/api/annotations"
-	math "math"
-)
 
-import (
 	context "context"
-	api "github.com/micro/go-micro/v2/api"
-	client "github.com/micro/go-micro/v2/client"
-	server "github.com/micro/go-micro/v2/server"
+
+	server "github.com/micro/micro/v3/service/server"
+
+	api "github.com/micro/micro/v3/service/api"
+
+	client "github.com/micro/micro/v3/service/client"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
