@@ -1,5 +1,5 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
-// source: micro/micro/cmd/usage/proto/usage.proto
+// source: cmd/usage/proto/usage.proto
 
 package usage
 
@@ -42,7 +42,7 @@ func (m *Usage) Reset()         { *m = Usage{} }
 func (m *Usage) String() string { return proto.CompactTextString(m) }
 func (*Usage) ProtoMessage()    {}
 func (*Usage) Descriptor() ([]byte, []int) {
-	return fileDescriptor_8a13186c7d202766, []int{0}
+	return fileDescriptor_d6c20769c80f20f4, []int{0}
 }
 
 func (m *Usage) XXX_Unmarshal(b []byte) error {
@@ -117,7 +117,7 @@ func (m *Metrics) Reset()         { *m = Metrics{} }
 func (m *Metrics) String() string { return proto.CompactTextString(m) }
 func (*Metrics) ProtoMessage()    {}
 func (*Metrics) Descriptor() ([]byte, []int) {
-	return fileDescriptor_8a13186c7d202766, []int{1}
+	return fileDescriptor_d6c20769c80f20f4, []int{1}
 }
 
 func (m *Metrics) XXX_Unmarshal(b []byte) error {
@@ -151,25 +151,23 @@ func init() {
 	proto.RegisterMapType((map[string]uint64)(nil), "Metrics.CountEntry")
 }
 
-func init() {
-	proto.RegisterFile("micro/micro/cmd/usage/proto/usage.proto", fileDescriptor_8a13186c7d202766)
-}
+func init() { proto.RegisterFile("cmd/usage/proto/usage.proto", fileDescriptor_d6c20769c80f20f4) }
 
-var fileDescriptor_8a13186c7d202766 = []byte{
-	// 238 bytes of a gzipped FileDescriptorProto
+var fileDescriptor_d6c20769c80f20f4 = []byte{
+	// 233 bytes of a gzipped FileDescriptorProto
 	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x44, 0x90, 0xcd, 0x4a, 0xc4, 0x30,
-	0x14, 0x85, 0x49, 0x7f, 0x9d, 0x3b, 0x20, 0x12, 0x45, 0x82, 0xb8, 0x28, 0xdd, 0x58, 0x37, 0x2d,
-	0x8c, 0x9b, 0xc1, 0xad, 0xb8, 0x74, 0x13, 0xf0, 0x01, 0x6a, 0x1a, 0xe4, 0xa2, 0x49, 0x86, 0x24,
-	0xed, 0x30, 0x0f, 0xe4, 0x7b, 0x4a, 0x93, 0x96, 0x6e, 0xc2, 0xf9, 0xbe, 0x73, 0x37, 0x27, 0xf0,
-	0xa4, 0x50, 0x58, 0xd3, 0xc5, 0x57, 0xa8, 0xa1, 0x1b, 0x5d, 0xff, 0x2d, 0xbb, 0x93, 0x35, 0xde,
-	0xc4, 0xdc, 0x86, 0x5c, 0xff, 0x11, 0xc8, 0x3f, 0x67, 0xa6, 0x0c, 0x4a, 0x27, 0xed, 0x84, 0x42,
-	0x32, 0x52, 0x91, 0x66, 0xc7, 0x57, 0x9c, 0x9b, 0x49, 0x5a, 0x87, 0x46, 0xb3, 0x24, 0x36, 0x0b,
-	0xd2, 0x6b, 0x48, 0x70, 0x60, 0x69, 0x90, 0x09, 0x0e, 0xf4, 0x11, 0x76, 0x1e, 0x95, 0x74, 0xbe,
-	0x57, 0x27, 0x96, 0x55, 0xa4, 0xc9, 0xf8, 0x26, 0xe8, 0x3d, 0x14, 0x67, 0xd4, 0x83, 0x39, 0xb3,
-	0x3c, 0x54, 0x0b, 0xd1, 0x1a, 0x4a, 0x25, 0xbd, 0x45, 0xe1, 0x58, 0x51, 0x91, 0x66, 0x7f, 0xb8,
-	0x6a, 0x3f, 0x22, 0xf3, 0xb5, 0xa8, 0x35, 0x94, 0x8b, 0xa3, 0xcf, 0x90, 0x0b, 0x33, 0x6a, 0xcf,
-	0x48, 0x95, 0x36, 0xfb, 0xc3, 0xed, 0x7a, 0xdc, 0xbe, 0xcd, 0xf6, 0x5d, 0x7b, 0x7b, 0xe1, 0xf1,
-	0xe2, 0xe1, 0x08, 0xb0, 0x49, 0x7a, 0x03, 0xe9, 0x8f, 0xbc, 0x2c, 0xeb, 0xe6, 0x48, 0xef, 0x20,
-	0x9f, 0xfa, 0xdf, 0x51, 0x86, 0x5d, 0x19, 0x8f, 0xf0, 0x9a, 0x1c, 0xc9, 0x57, 0x11, 0xbe, 0xe7,
-	0xe5, 0x3f, 0x00, 0x00, 0xff, 0xff, 0x11, 0x9b, 0xc4, 0xfb, 0x49, 0x01, 0x00, 0x00,
+	0x14, 0x85, 0x49, 0x7f, 0x9d, 0x3b, 0x20, 0x12, 0x45, 0x2e, 0xea, 0xa2, 0x74, 0x55, 0x37, 0x1d,
+	0x18, 0x37, 0x83, 0x5b, 0x71, 0xe9, 0x26, 0xe0, 0x03, 0xd4, 0xf6, 0x22, 0x41, 0x9b, 0x0c, 0x49,
+	0xda, 0x61, 0x1e, 0xc8, 0xf7, 0x94, 0x26, 0x29, 0xb3, 0x3b, 0xdf, 0x77, 0xee, 0xe6, 0x5c, 0x78,
+	0xec, 0xc7, 0x61, 0x37, 0xd9, 0xee, 0x9b, 0x76, 0x47, 0xa3, 0x9d, 0x0e, 0xb9, 0xf5, 0xb9, 0xfe,
+	0x63, 0x90, 0x7f, 0x2e, 0xcc, 0x11, 0x4a, 0x4b, 0x66, 0x96, 0x3d, 0x21, 0xab, 0x58, 0xb3, 0x11,
+	0x2b, 0x2e, 0xcd, 0x4c, 0xc6, 0x4a, 0xad, 0x30, 0x09, 0x4d, 0x44, 0x7e, 0x0d, 0x89, 0x1c, 0x30,
+	0xf5, 0x32, 0x91, 0x03, 0x7f, 0x82, 0x8d, 0x93, 0x23, 0x59, 0xd7, 0x8d, 0x47, 0xcc, 0x2a, 0xd6,
+	0x64, 0xe2, 0x22, 0xf8, 0x3d, 0x14, 0x27, 0xa9, 0x06, 0x7d, 0xc2, 0xdc, 0x57, 0x91, 0x78, 0x0d,
+	0xe5, 0x48, 0xce, 0xc8, 0xde, 0x62, 0x51, 0xb1, 0x66, 0xbb, 0xbf, 0x6a, 0x3f, 0x02, 0x8b, 0xb5,
+	0xa8, 0x15, 0x94, 0xd1, 0xf1, 0x67, 0xc8, 0x7b, 0x3d, 0x29, 0x87, 0xac, 0x4a, 0x9b, 0xed, 0xfe,
+	0x76, 0x3d, 0x6e, 0xdf, 0x16, 0xfb, 0xae, 0x9c, 0x39, 0x8b, 0x70, 0xf1, 0x70, 0x00, 0xb8, 0x48,
+	0x7e, 0x03, 0xe9, 0x0f, 0x9d, 0xe3, 0xba, 0x25, 0xf2, 0x3b, 0xc8, 0xe7, 0xee, 0x77, 0x22, 0xbf,
+	0x2b, 0x13, 0x01, 0x5e, 0x93, 0x03, 0xfb, 0x2a, 0xfc, 0x7b, 0x5e, 0xfe, 0x03, 0x00, 0x00, 0xff,
+	0xff, 0xe4, 0x89, 0x4c, 0x0e, 0x3d, 0x01, 0x00, 0x00,
 }
