@@ -1,30 +1,22 @@
 package auth
 
-import "github.com/micro/go-micro/v2/auth"
+import (
+	"github.com/micro/micro/v3/service/auth"
+)
 
-// TokenCookieName is the name of the cookie which stores the auth token
-const TokenCookieName = "micro-token"
+const (
+	// BearerScheme used for Authorization header
+	BearerScheme = "Bearer "
+	// TokenCookieName is the name of the cookie which stores the auth token
+	TokenCookieName = "micro-token"
+)
 
 // SystemRules are the default rules which are applied to the runtime services
 var SystemRules = []*auth.Rule{
 	&auth.Rule{
 		ID:       "default",
-		Scope:    "*",
+		Scope:    auth.ScopePublic,
+		Access:   auth.AccessGranted,
 		Resource: &auth.Resource{Type: "*", Name: "*", Endpoint: "*"},
-	},
-	&auth.Rule{
-		ID:       "auth-public",
-		Scope:    "",
-		Resource: &auth.Resource{Type: "service", Name: "go.micro.auth", Endpoint: "*"},
-	},
-	&auth.Rule{
-		ID:       "registry-get",
-		Scope:    "",
-		Resource: &auth.Resource{Type: "service", Name: "go.micro.registry", Endpoint: "Registry.GetService"},
-	},
-	&auth.Rule{
-		ID:       "registry-list",
-		Scope:    "",
-		Resource: &auth.Resource{Type: "service", Name: "go.micro.registry", Endpoint: "Registry.ListServices"},
 	},
 }
