@@ -5,12 +5,13 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"errors"
+	"fmt"
 	"io/ioutil"
 	"net/http"
 	"strings"
 
-	"github.com/micro/cli/v2"
-	"github.com/micro/go-micro/v2/metadata"
+	"github.com/micro/micro/v3/service/context/metadata"
+	"github.com/urfave/cli/v2"
 )
 
 func ACMEHosts(ctx *cli.Context) []string {
@@ -66,4 +67,27 @@ func TLSConfig(ctx *cli.Context) (*tls.Config, error) {
 	}
 
 	return nil, errors.New("TLS certificate and key files not specified")
+}
+
+// UnexpectedSubcommand checks for erroneous subcommands and prints help and returns error
+func UnexpectedSubcommand(ctx *cli.Context) error {
+	if first := Subcommand(ctx); first != "" {
+		// received something that isn't a subcommand
+		return cli.Exit(fmt.Sprintf("Unrecognized subcommand for %s: %s. Please refer to '%s --help'", ctx.App.Name, first, ctx.App.Name), 1)
+	}
+	return cli.ShowSubcommandHelp(ctx)
+}
+
+func UnexpectedCommand(ctx *cli.Context) error {
+	commandName := ctx.Args().First()
+	return cli.Exit(fmt.Sprintf("Unrecognized micro command: %s. Please refer to 'micro --help'", commandName), 1)
+}
+
+func MissingCommand(ctx *cli.Context) error {
+	return cli.Exit(fmt.Sprintf("No command provided to micro. Please refer to 'micro --help'"), 1)
+}
+
+// MicroSubcommand returns the subcommand name
+func Subcommand(ctx *cli.Context) string {
+	return ctx.Args().First()
 }
