@@ -4,7 +4,8 @@ package micro
 import (
 	"net/http"
 
-	"github.com/micro/go-micro/v2/api/resolver"
+	"github.com/micro/micro/v3/internal/api/resolver"
+	"github.com/micro/micro/v3/service/registry"
 )
 
 // default resolver for legacy purposes
@@ -12,13 +13,15 @@ import (
 // /foo becomes namespace.foo
 // /v1/foo becomes namespace.v1.foo
 type Resolver struct {
-	Options resolver.Options
+	opts resolver.Options
 }
 
-func (r *Resolver) Resolve(req *http.Request) (*resolver.Endpoint, error) {
+func (r *Resolver) Resolve(req *http.Request, opts ...resolver.ResolveOption) (*resolver.Endpoint, error) {
+	options := resolver.NewResolveOptions(opts...)
+
 	var name, method string
 
-	switch r.Options.Handler {
+	switch r.opts.Handler {
 	// internal handlers
 	case "meta", "api", "rpc", "micro":
 		name, method = apiRoute(req.URL.Path)
@@ -27,9 +30,24 @@ func (r *Resolver) Resolve(req *http.Request) (*resolver.Endpoint, error) {
 		name = proxyRoute(req.URL.Path)
 	}
 
-	ns := r.Options.Namespace(req)
+	// append the service prefix, e.g. foo.api
+	if len(r.opts.ServicePrefix) > 0 {
+		name = r.opts.ServicePrefix + "." + name
+	}
+
+	// check for the namespace in the request header, this can be set by the client or injected
+	// by the auth wrapper if an auth token was provided. The headr takes priority over any domain
+	// passed as a default
+	domain := options.Domain
+	if dom := req.Header.Get("Micro-Namespace"); len(dom) > 0 && dom != domain {
+		domain = dom
+	} else if len(domain) == 0 {
+		domain = registry.DefaultDomain
+	}
+
 	return &resolver.Endpoint{
-		Name:   ns + "." + name,
+		Name:   name,
+		Domain: domain,
 		Method: method,
 	}, nil
 }
@@ -41,6 +59,6 @@ func (r *Resolver) String() string {
 // NewResolver creates a new micro resolver
 func NewResolver(opts ...resolver.Option) resolver.Resolver {
 	return &Resolver{
-		Options: resolver.NewOptions(opts...),
+		opts: resolver.NewOptions(opts...),
 	}
 }
