@@ -3,23 +3,18 @@ package template
 var (
 	Makefile = `
 GOPATH:=$(shell go env GOPATH)
-MODIFY=Mgithub.com/micro/go-micro/api/proto/api.proto=github.com/micro/go-micro/v2/api/proto
-{{if ne .Type "web"}}
+.PHONY: init
+init:
+	go get -u github.com/golang/protobuf/proto
+	go get -u github.com/golang/protobuf/protoc-gen-go
+	go get github.com/micro/micro/v3/cmd/protoc-gen-micro
 .PHONY: proto
 proto:
-    {{if eq .UseGoPath true}}
-	protoc --proto_path=${GOPATH}/src:. --micro_out=${MODIFY}:. --go_out=${MODIFY}. proto/{{.Alias}}/{{.Alias}}.proto
-    {{else}}
-	protoc --proto_path=. --micro_out=${MODIFY}:. --go_out=${MODIFY}:. proto/{{.Alias}}/{{.Alias}}.proto
-    {{end}}
-
-.PHONY: build
-build: proto
-{{else}}
+	protoc --proto_path=. --micro_out=. --go_out=:. proto/{{.Alias}}.proto
+	
 .PHONY: build
 build:
-{{end}}
-	go build -o {{.Alias}}-{{.Type}} *.go
+	go build -o {{.Alias}} *.go
 
 .PHONY: test
 test:
@@ -27,7 +22,7 @@ test:
 
 .PHONY: docker
 docker:
-	docker build . -t {{.Alias}}-{{.Type}}:latest
+	docker build . -t {{.Alias}}:latest
 `
 
 	GenerateFile = `package main
