@@ -3,7 +3,11 @@ package template
 var (
 	Module = `module {{.Dir}}
 
-go 1.13
+go 1.15
+
+require (
+	github.com/micro/micro/v3 v3.0.0
+)
 
 // This can be removed once etcd becomes go gettable, version 3.4 and 3.5 is not,
 // see https://github.com/etcd-io/etcd/issues/11154 and https://github.com/etcd-io/etcd/issues/11931.
