@@ -1,32 +1,11 @@
 package template
 
 var (
-	ProtoFNC = `syntax = "proto3";
-
-package {{.FQDN}};
-
-option go_package = "{{.Dir}}/proto/{{.Alias}}";
-
-service {{title .Alias}} {
-	rpc Call(Request) returns (Response) {}
-}
-
-message Message {
-	string say = 1;
-}
-
-message Request {
-	string name = 1;
-}
-
-message Response {
-	string msg = 1;
-}
-`
-
 	ProtoSRV = `syntax = "proto3";
 
-package {{.FQDN}};
+package {{dehyphen .Alias}};
+
+option go_package = "proto;{{dehyphen .Alias}}";
 
 service {{title .Alias}} {
 	rpc Call(Request) returns (Response) {}
@@ -60,17 +39,6 @@ message Ping {
 
 message Pong {
 	int64 stroke = 1;
-}
-`
-
-	ProtoAPI = `syntax = "proto3";
-
-package {{.FQDN}};
-
-import "github.com/micro/go-micro/api/proto/api.proto";
-
-service {{title .Alias}} {
-	rpc Call(go.api.Request) returns (go.api.Response) {}
 }
 `
 )
