@@ -7,10 +7,10 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/micro/cli/v2"
-	"github.com/micro/go-micro/v2/registry"
-	"github.com/micro/go-micro/v2/util/backoff"
-	"github.com/micro/micro/v2/plugin"
+	"github.com/micro/micro/v3/internal/backoff"
+	"github.com/micro/micro/v3/plugin"
+	"github.com/micro/micro/v3/service/registry"
+	"github.com/urfave/cli/v2"
 )
 
 func init() {
@@ -38,6 +38,14 @@ func Plugin() plugin.Plugin {
 			// set service name
 			if c.Args().Len() > 0 && len(c.Args().Get(0)) > 0 {
 				service = c.Args().Get(0)
+			}
+
+			// service subcommand
+			if service == "service" {
+				// set as the sub command
+				if v := c.Args().Get(1); len(v) > 0 {
+					service = v
+				}
 			}
 
 			// kick off the tracker
