@@ -1,11 +1,11 @@
 // Code generated by protoc-gen-micro. DO NOT EDIT.
-// source: github.com/micro/micro/v2/internal/usage/proto/usage.proto
+// source: github.com/micro/micro/v3/internal/usage/proto/usage.proto
 
 /*
 Package usage is a generated protocol buffer package.
 
 It is generated from these files:
-	github.com/micro/micro/v2/internal/usage/proto/usage.proto
+	github.com/micro/micro/v3/internal/usage/proto/usage.proto
 
 It has these top-level messages:
 	Usage
