@@ -1,11 +1,13 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
-// source: github.com/micro/micro/v2/internal/usage/proto/usage.proto
+// source: internal/usage/proto/usage.proto
 
 package usage
 
-import proto "github.com/golang/protobuf/proto"
-import fmt "fmt"
-import math "math"
+import (
+	fmt "fmt"
+	proto "github.com/golang/protobuf/proto"
+	math "math"
+)
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ = proto.Marshal
@@ -16,7 +18,7 @@ var _ = math.Inf
 // is compatible with the proto package it is being compiled against.
 // A compilation error at this line likely means your copy of the
 // proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion2 // please upgrade the proto package
+const _ = proto.ProtoPackageIsVersion3 // please upgrade the proto package
 
 type Usage struct {
 	// service name
@@ -40,16 +42,17 @@ func (m *Usage) Reset()         { *m = Usage{} }
 func (m *Usage) String() string { return proto.CompactTextString(m) }
 func (*Usage) ProtoMessage()    {}
 func (*Usage) Descriptor() ([]byte, []int) {
-	return fileDescriptor_usage_7cdccb5f3bba2052, []int{0}
+	return fileDescriptor_57da1bcc49fd7460, []int{0}
 }
+
 func (m *Usage) XXX_Unmarshal(b []byte) error {
 	return xxx_messageInfo_Usage.Unmarshal(m, b)
 }
 func (m *Usage) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
 	return xxx_messageInfo_Usage.Marshal(b, m, deterministic)
 }
-func (dst *Usage) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Usage.Merge(dst, src)
+func (m *Usage) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_Usage.Merge(m, src)
 }
 func (m *Usage) XXX_Size() int {
 	return xxx_messageInfo_Usage.Size(m)
@@ -114,16 +117,17 @@ func (m *Metrics) Reset()         { *m = Metrics{} }
 func (m *Metrics) String() string { return proto.CompactTextString(m) }
 func (*Metrics) ProtoMessage()    {}
 func (*Metrics) Descriptor() ([]byte, []int) {
-	return fileDescriptor_usage_7cdccb5f3bba2052, []int{1}
+	return fileDescriptor_57da1bcc49fd7460, []int{1}
 }
+
 func (m *Metrics) XXX_Unmarshal(b []byte) error {
 	return xxx_messageInfo_Metrics.Unmarshal(m, b)
 }
 func (m *Metrics) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
 	return xxx_messageInfo_Metrics.Marshal(b, m, deterministic)
 }
-func (dst *Metrics) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_Metrics.Merge(dst, src)
+func (m *Metrics) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_Metrics.Merge(m, src)
 }
 func (m *Metrics) XXX_Size() int {
 	return xxx_messageInfo_Metrics.Size(m)
@@ -147,26 +151,23 @@ func init() {
 	proto.RegisterMapType((map[string]uint64)(nil), "Metrics.CountEntry")
 }
 
-func init() {
-	proto.RegisterFile("github.com/micro/micro/v2/internal/usage/proto/usage.proto", fileDescriptor_usage_7cdccb5f3bba2052)
-}
+func init() { proto.RegisterFile("internal/usage/proto/usage.proto", fileDescriptor_57da1bcc49fd7460) }
 
-var fileDescriptor_usage_7cdccb5f3bba2052 = []byte{
-	// 251 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x44, 0x90, 0xbd, 0x6a, 0xc3, 0x30,
-	0x14, 0x85, 0x91, 0xe3, 0x9f, 0xe6, 0x06, 0x4a, 0x51, 0x4b, 0x11, 0xa5, 0x83, 0xf1, 0xe4, 0x2e,
-	0x36, 0xa4, 0x43, 0x43, 0xd7, 0xd2, 0xb1, 0x8b, 0xa1, 0x0f, 0xe0, 0xc8, 0x22, 0xbd, 0x34, 0x92,
-	0x82, 0x24, 0x3b, 0xe4, 0x81, 0xfa, 0x9e, 0xc5, 0x92, 0x4c, 0x16, 0x71, 0xbe, 0xef, 0xdc, 0xe5,
-	0x08, 0xde, 0x0e, 0xe8, 0x7e, 0xc6, 0x7d, 0xc3, 0xb5, 0x6c, 0x25, 0x72, 0xa3, 0xe3, 0x8b, 0xca,
-	0x09, 0xa3, 0xfa, 0x63, 0x3b, 0xda, 0xfe, 0x20, 0xda, 0x93, 0xd1, 0x4e, 0x87, 0xdc, 0xf8, 0x5c,
-	0xfd, 0x11, 0xc8, 0xbe, 0x67, 0xa6, 0x0c, 0x0a, 0x2b, 0xcc, 0x84, 0x5c, 0x30, 0x52, 0x92, 0x7a,
-	0xdd, 0x2d, 0x38, 0x37, 0x93, 0x30, 0x16, 0xb5, 0x62, 0x49, 0x68, 0x22, 0xd2, 0x5b, 0x48, 0x70,
-	0x60, 0x2b, 0x2f, 0x13, 0x1c, 0xe8, 0x33, 0xac, 0x1d, 0x4a, 0x61, 0x5d, 0x2f, 0x4f, 0x2c, 0x2d,
-	0x49, 0x9d, 0x76, 0x57, 0x41, 0x1f, 0x21, 0x3f, 0xa3, 0x1a, 0xf4, 0x99, 0x65, 0xbe, 0x8a, 0x44,
-	0x2b, 0x28, 0xa4, 0x70, 0x06, 0xb9, 0x65, 0x79, 0x49, 0xea, 0xcd, 0xf6, 0xa6, 0xf9, 0x0a, 0xdc,
-	0x2d, 0x45, 0xa5, 0xa0, 0x88, 0x8e, 0xbe, 0x40, 0xc6, 0xf5, 0xa8, 0x1c, 0x23, 0xe5, 0xaa, 0xde,
-	0x6c, 0xef, 0x97, 0xe3, 0xe6, 0x63, 0xb6, 0x9f, 0xca, 0x99, 0x4b, 0x17, 0x2e, 0x9e, 0x76, 0x00,
-	0x57, 0x49, 0xef, 0x60, 0xf5, 0x2b, 0x2e, 0x71, 0xdd, 0x1c, 0xe9, 0x03, 0x64, 0x53, 0x7f, 0x1c,
-	0x85, 0xdf, 0x95, 0x76, 0x01, 0xde, 0x93, 0x1d, 0xd9, 0xe7, 0xfe, 0x7b, 0x5e, 0xff, 0x03, 0x00,
-	0x00, 0xff, 0xff, 0x4f, 0xfc, 0xa1, 0x6c, 0x59, 0x01, 0x00, 0x00,
+var fileDescriptor_57da1bcc49fd7460 = []byte{
+	// 236 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x44, 0x90, 0xcd, 0x4a, 0xc4, 0x30,
+	0x14, 0x85, 0x49, 0x7f, 0x9d, 0x3b, 0x20, 0x12, 0x45, 0x82, 0xb8, 0x08, 0x5d, 0xd5, 0x4d, 0x07,
+	0xc6, 0xcd, 0xe0, 0x56, 0x5c, 0xba, 0x09, 0xf8, 0x00, 0xb5, 0xbd, 0xc8, 0xc5, 0x69, 0x32, 0x24,
+	0x69, 0x87, 0x79, 0x20, 0xdf, 0x53, 0x9a, 0xb4, 0xcc, 0xee, 0x7c, 0xdf, 0xb9, 0x9b, 0x73, 0x41,
+	0x92, 0xf6, 0x68, 0x75, 0x7b, 0xdc, 0x8d, 0xae, 0xfd, 0xc1, 0xdd, 0xc9, 0x1a, 0x6f, 0x62, 0x6e,
+	0x42, 0xae, 0xfe, 0x18, 0xe4, 0x5f, 0x33, 0x73, 0x01, 0xa5, 0x43, 0x3b, 0x51, 0x87, 0x82, 0x49,
+	0x56, 0x6f, 0xd4, 0x8a, 0x73, 0x33, 0xa1, 0x75, 0x64, 0xb4, 0x48, 0x62, 0xb3, 0x20, 0xbf, 0x85,
+	0x84, 0x7a, 0x91, 0x06, 0x99, 0x50, 0xcf, 0x9f, 0x61, 0xe3, 0x69, 0x40, 0xe7, 0xdb, 0xe1, 0x24,
+	0x32, 0xc9, 0xea, 0x4c, 0x5d, 0x05, 0x7f, 0x84, 0xe2, 0x4c, 0xba, 0x37, 0x67, 0x91, 0x87, 0x6a,
+	0x21, 0x5e, 0x41, 0x39, 0xa0, 0xb7, 0xd4, 0x39, 0x51, 0x48, 0x56, 0x6f, 0xf7, 0x37, 0xcd, 0x67,
+	0x64, 0xb5, 0x16, 0x95, 0x86, 0x72, 0x71, 0xfc, 0x05, 0xf2, 0xce, 0x8c, 0xda, 0x0b, 0x26, 0xd3,
+	0x7a, 0xbb, 0xbf, 0x5f, 0x8f, 0x9b, 0xf7, 0xd9, 0x7e, 0x68, 0x6f, 0x2f, 0x2a, 0x5e, 0x3c, 0x1d,
+	0x00, 0xae, 0x92, 0xdf, 0x41, 0xfa, 0x8b, 0x97, 0x65, 0xdd, 0x1c, 0xf9, 0x03, 0xe4, 0x53, 0x7b,
+	0x1c, 0x31, 0xec, 0xca, 0x54, 0x84, 0xb7, 0xe4, 0xc0, 0xbe, 0x8b, 0xf0, 0x9e, 0xd7, 0xff, 0x00,
+	0x00, 0x00, 0xff, 0xff, 0x8d, 0x2b, 0xa5, 0xb9, 0x42, 0x01, 0x00, 0x00,
 }
