@@ -5,10 +5,12 @@ import (
 	"sync"
 )
 
+const defaultModule = "micro"
+
 type manager struct {
 	sync.Mutex
-	plugins    []Plugin
-	registered map[string]bool
+	plugins    map[string][]Plugin
+	registered map[string]map[string]bool
 }
 
 var (
@@ -18,34 +20,68 @@ var (
 
 func newManager() *manager {
 	return &manager{
-		registered: make(map[string]bool),
+		plugins:    make(map[string][]Plugin),
+		registered: make(map[string]map[string]bool),
 	}
 }
 
-func (m *manager) Plugins() []Plugin {
+func (m *manager) Plugins(opts ...PluginOption) []Plugin {
+	options := PluginOptions{Module: defaultModule}
+	for _, o := range opts {
+		o(&options)
+	}
+
 	m.Lock()
 	defer m.Unlock()
-	return m.plugins
+
+	if plugins, ok := m.plugins[options.Module]; ok {
+		return plugins
+	}
+	return []Plugin{}
 }
 
-func (m *manager) Register(plugin Plugin) error {
+func (m *manager) Register(plugin Plugin, opts ...PluginOption) error {
+	options := PluginOptions{Module: defaultModule}
+	for _, o := range opts {
+		o(&options)
+	}
+
 	m.Lock()
 	defer m.Unlock()
 
 	name := plugin.String()
 
-	if m.registered[name] {
+	if reg, ok := m.registered[options.Module]; ok && reg[name] {
 		return fmt.Errorf("Plugin with name %s already registered", name)
 	}
 
-	m.registered[name] = true
-	m.plugins = append(m.plugins, plugin)
+	if _, ok := m.registered[options.Module]; !ok {
+		m.registered[options.Module] = map[string]bool{name: true}
+	} else {
+		m.registered[options.Module][name] = true
+	}
+
+	if _, ok := m.plugins[options.Module]; ok {
+		m.plugins[options.Module] = append(m.plugins[options.Module], plugin)
+	} else {
+		m.plugins[options.Module] = []Plugin{plugin}
+	}
+
 	return nil
 }
 
-func (m *manager) isRegistered(plugin Plugin) bool {
+func (m *manager) isRegistered(plugin Plugin, opts ...PluginOption) bool {
+	options := PluginOptions{Module: defaultModule}
+	for _, o := range opts {
+		o(&options)
+	}
+
 	m.Lock()
 	defer m.Unlock()
 
-	return m.registered[plugin.String()]
+	if _, ok := m.registered[options.Module]; !ok {
+		return false
+	}
+
+	return m.registered[options.Module][plugin.String()]
 }
