@@ -1,7 +1,7 @@
 package plugin
 
 import (
-	"github.com/micro/cli/v2"
+	"github.com/urfave/cli/v2"
 )
 
 // Options are used as part of a new plugin
