@@ -1,229 +1,75 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Original source: github.com/micro/go-micro/v3/config/config.go
+
+// Package config is an interface for dynamic configuration.
 package config
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"strings"
 	"time"
-
-	"github.com/micro/cli/v2"
-	"github.com/micro/go-micro/v2"
-	"github.com/micro/go-micro/v2/config/cmd"
-	proto "github.com/micro/go-micro/v2/config/source/service/proto"
-	log "github.com/micro/go-micro/v2/logger"
-	"github.com/micro/micro/v2/internal/client"
-	"github.com/micro/micro/v2/service/config/handler"
 )
 
-var (
-	// Service name
-	Name = "go.micro.config"
-	// Default database store
-	Database = "store"
-	// Default key
-	Namespace = "global"
-)
+// DefaultConfig implementation
+var DefaultConfig Config
 
-func Run(c *cli.Context, srvOpts ...micro.Option) {
-	if len(c.String("server_name")) > 0 {
-		Name = c.String("server_name")
-	}
+// Config is an interface abstraction for dynamic configuration
+type Config interface {
+	Get(path string, options ...Option) (Value, error)
+	Set(path string, val interface{}, options ...Option) error
+	Delete(path string, options ...Option) error
+}
 
-	if len(c.String("watch_topic")) > 0 {
-		handler.WatchTopic = c.String("watch_topic")
-	}
+// Value represents a value of any type
+type Value interface {
+	Exists() bool
+	Bool(def bool) bool
+	Int(def int) int
+	String(def string) string
+	Float64(def float64) float64
+	Duration(def time.Duration) time.Duration
+	StringSlice(def []string) []string
+	StringMap(def map[string]string) map[string]string
+	Scan(val interface{}) error
+	Bytes() []byte
+}
 
-	srvOpts = append(srvOpts, micro.Name(Name))
+type Options struct {
+	Secret bool
+}
 
-	service := micro.NewService(srvOpts...)
+type Option func(o *Options)
 
-	h := &handler.Config{
-		Store: *cmd.DefaultCmd.Options().Store,
-	}
-
-	proto.RegisterConfigHandler(service.Server(), h)
-	micro.RegisterSubscriber(handler.WatchTopic, service.Server(), handler.Watcher)
-
-	if err := service.Run(); err != nil {
-		log.Fatalf("config Run the service error: ", err)
+func Secret(b bool) Option {
+	return func(o *Options) {
+		o.Secret = b
 	}
 }
 
-func setConfig(ctx *cli.Context) error {
-	pb := proto.NewConfigService("go.micro.config", client.New(ctx))
-
-	args := ctx.Args()
-
-	if args.Len() == 0 {
-		fmt.Println("Required usage: micro config set key val")
-		os.Exit(1)
-	}
-
-	// key val
-	key := args.Get(0)
-	val := args.Get(1)
-
-	// TODO: allow the specifying of a config.Key. This will be service name
-	// The actuall key-val set is a path e.g micro/accounts/key
-	_, err := pb.Update(context.TODO(), &proto.UpdateRequest{
-		Change: &proto.Change{
-			// global key
-			Namespace: Namespace,
-			// actual key for the value
-			Path: key,
-			// The value
-			ChangeSet: &proto.ChangeSet{
-				Data:      string(val),
-				Format:    "json",
-				Source:    "cli",
-				Timestamp: time.Now().Unix(),
-			},
-		},
-	})
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-
-	return nil
+type Secrets interface {
+	Config
 }
 
-func getConfig(ctx *cli.Context) error {
-	pb := proto.NewConfigService("go.micro.config", client.New(ctx))
-
-	args := ctx.Args()
-
-	if args.Len() == 0 {
-		fmt.Println("Required usage: micro config get key")
-		os.Exit(1)
-	}
-
-	// key val
-	key := args.Get(0)
-
-	if len(key) == 0 {
-		log.Fatal("key cannot be blank")
-	}
-
-	// TODO: allow the specifying of a config.Key. This will be service name
-	// The actuall key-val set is a path e.g micro/accounts/key
-
-	rsp, err := pb.Read(context.TODO(), &proto.ReadRequest{
-		// The global key,
-		Namespace: Namespace,
-		// The actual key for the val
-		Path: key,
-	})
-	if err != nil {
-		if strings.Contains(err.Error(), "not found") {
-			fmt.Println("not found")
-			os.Exit(1)
-		}
-		fmt.Println(err)
-		os.Exit(1)
-	}
-
-	if rsp.Change == nil || rsp.Change.ChangeSet == nil {
-		fmt.Println("not found")
-		os.Exit(1)
-	}
-
-	// don't do it
-	if v := rsp.Change.ChangeSet.Data; len(v) == 0 || string(v) == "null" {
-		fmt.Println("not found")
-		os.Exit(1)
-	}
-
-	fmt.Println(string(rsp.Change.ChangeSet.Data))
-
-	return nil
+// Get a value at the path
+func Get(path string, options ...Option) (Value, error) {
+	return DefaultConfig.Get(path, options...)
 }
 
-func delConfig(ctx *cli.Context) error {
-	pb := proto.NewConfigService("go.micro.config", client.New(ctx))
-
-	args := ctx.Args()
-
-	if args.Len() == 0 {
-		fmt.Println("Required usage: micro config get key")
-		os.Exit(1)
-	}
-
-	// key val
-	key := args.Get(0)
-
-	if len(key) == 0 {
-		log.Fatal("key cannot be blank")
-	}
-
-	// TODO: allow the specifying of a config.Key. This will be service name
-	// The actuall key-val set is a path e.g micro/accounts/key
-
-	_, err := pb.Delete(context.TODO(), &proto.DeleteRequest{
-		Change: &proto.Change{
-			// The global key,
-			Namespace: Namespace,
-			// The actual key for the val
-			Path: key,
-		},
-	})
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-
-	return nil
+// Set the value at a path
+func Set(path string, val interface{}, options ...Option) error {
+	return DefaultConfig.Set(path, val, options...)
 }
 
-func Commands(options ...micro.Option) []*cli.Command {
-	command := &cli.Command{
-		Name:  "config",
-		Usage: "Manage configuration values",
-		Subcommands: []*cli.Command{
-			{
-				Name:   "set",
-				Usage:  "Set a key-val; micro config set key val",
-				Action: setConfig,
-			},
-			{
-				Name:   "get",
-				Usage:  "Get a value; micro config get key",
-				Action: getConfig,
-			},
-			{
-				Name:   "del",
-				Usage:  "Delete a value; micro config del key",
-				Action: delConfig,
-			},
-		},
-		Action: func(ctx *cli.Context) error {
-			Run(ctx, options...)
-			return nil
-		},
-		Flags: []cli.Flag{
-			&cli.StringFlag{
-				Name:    "namespace",
-				EnvVars: []string{"MICRO_CONFIG_NAMESPACE"},
-				Usage:   "Set the namespace used by the Config Service e.g. go.micro.srv.config",
-			},
-			&cli.StringFlag{
-				Name:    "watch_topic",
-				EnvVars: []string{"MICRO_CONFIG_WATCH_TOPIC"},
-				Usage:   "watch the change event.",
-			},
-		},
-	}
-
-	for _, p := range Plugins() {
-		if cmds := p.Commands(); len(cmds) > 0 {
-			command.Subcommands = append(command.Subcommands, cmds...)
-		}
-
-		if flags := p.Flags(); len(flags) > 0 {
-			command.Flags = append(command.Flags, flags...)
-		}
-	}
-
-	return []*cli.Command{command}
+// Delete the value at a path
+func Delete(path string, options ...Option) error {
+	return DefaultConfig.Delete(path, options...)
 }
