@@ -9,10 +9,10 @@ import (
 )
 
 func TestEnvBasic(t *testing.T) {
-	trySuite(t, testEnvOverrides, retryCount)
+	TrySuite(t, testEnvOverrides, retryCount)
 }
 
-func testEnvBasic(t *t) {
+func testEnvBasic(t *T) {
 	outp, err := exec.Command("micro", "-env=platform", "env").CombinedOutput()
 	if err != nil {
 		t.Fatal(err)
@@ -28,10 +28,10 @@ func testEnvBasic(t *t) {
 }
 
 func TestEnvOverrides(t *testing.T) {
-	trySuite(t, testEnvOverrides, retryCount)
+	TrySuite(t, testEnvOverrides, retryCount)
 }
 
-func testEnvOverrides(t *t) {
+func testEnvOverrides(t *T) {
 	outp, err := exec.Command("micro", "-env=platform", "env").CombinedOutput()
 	if err != nil {
 		t.Fatal(err)
@@ -51,4 +51,66 @@ func testEnvOverrides(t *t) {
 		t.Fatal("Env platform is not selected")
 		return
 	}
+}
+
+func TestEnvOps(t *testing.T) {
+	TrySuite(t, testEnvOps, retryCount)
+}
+
+func testEnvOps(t *T) {
+	// add an env
+	_, err := exec.Command("micro", "env", "add", "fooTestEnvOps", "127.0.0.1:8081").CombinedOutput()
+	if err != nil {
+		t.Fatalf("Error running micro env %s", err)
+		return
+	}
+	outp, err := exec.Command("micro", "env").CombinedOutput()
+	if err != nil {
+		t.Fatalf("Error running micro env %s", err)
+		return
+	}
+	if !strings.Contains(string(outp), "fooTestEnvOps") {
+		t.Fatalf("Cannot find expected environment. Output %s", outp)
+		return
+	}
+
+	// can we actually set it correctly
+	_, err = exec.Command("micro", "env", "set", "fooTestEnvOps").CombinedOutput()
+	if err != nil {
+		t.Fatalf("Error running micro env %s", err)
+		return
+	}
+	outp, err = exec.Command("micro", "env").CombinedOutput()
+	if err != nil {
+		t.Fatalf("Error running micro env %s", err)
+		return
+	}
+	if !strings.Contains(string(outp), "* fooTestEnvOps") {
+		t.Fatalf("Environment not set. Output %s", outp)
+		return
+	}
+
+	_, err = exec.Command("micro", "env", "set", "local").CombinedOutput()
+	if err != nil {
+		t.Fatalf("Error running micro env %s", err)
+		return
+	}
+
+	// we should be able to delete it too
+	_, err = exec.Command("micro", "env", "del", "fooTestEnvOps").CombinedOutput()
+	if err != nil {
+		t.Fatalf("Error running micro env %s", err)
+		return
+	}
+
+	outp, err = exec.Command("micro", "env").CombinedOutput()
+	if err != nil {
+		t.Fatalf("Error running micro env %s", err)
+		return
+	}
+	if strings.Contains(string(outp), "fooTestEnvOps") {
+		t.Fatalf("Found unexpected environment. Output %s", outp)
+		return
+	}
+
 }
